requirements.txt
cryptography>=41.0.0
argon2-cffi>=23.1.0
pynput>=1.7.6
pyperclip>=1.8.2
Pillow>=10.0.0
pystray>=0.19.5
secure-smtplib>=0.1.1
pytest>=7.4.0
keyring>=24.0.0
customtkinter>=5.2.2
